requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg==3.1.10
PyJWT==2.8.0
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.7
